requirements.txt
selenium==4.4.3
webdriver-manager==3.8.4
pytest==7.1.2
allure-pytest==2.10.0
allure-python-commons==2.10.0
pytest-xdist==2.5.0
requests==2.28.1
jsonschema~=4.7.2
mimesis~=5.5.0
